agent/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
supabase>=1.0.3
langchain>=0.0.325
langgraph>=0.0.15 
langchain-openai>=0.0.325